Switch all non-generated Silicon Labs that we use right now to C++ (#3581)
