+++ examples/lock-app/efr32/BUILD.gn
@@ -86,18 +86,18 @@ efr32_executable("lock_app") {
   sources = [
     "${chip_root}/examples/common/QRCode/repo/c/qrcodegen.c",
     "${chip_root}/examples/common/chip-app-server/DataModelHandler.cpp",
-    "${chip_root}/src/app/clusters/on-off-server/on-off.c",
+    "${chip_root}/src/app/clusters/on-off-server/on-off.cpp",
     "${chip_root}/src/app/util/af-event.cpp",
     "${chip_root}/src/app/util/af-main-common.cpp",
-    "${chip_root}/src/app/util/attribute-size.c",
-    "${chip_root}/src/app/util/attribute-storage.c",
-    "${chip_root}/src/app/util/attribute-table.c",
+    "${chip_root}/src/app/util/attribute-size.cpp",
+    "${chip_root}/src/app/util/attribute-storage.cpp",
+    "${chip_root}/src/app/util/attribute-table.cpp",
     "${chip_root}/src/app/util/binding-table.cpp",
     "${chip_root}/src/app/util/chip-message-send.cpp",
-    "${chip_root}/src/app/util/client-api.c",
+    "${chip_root}/src/app/util/client-api.cpp",
     "${chip_root}/src/app/util/ember-print.cpp",
     "${chip_root}/src/app/util/message.cpp",
-    "${chip_root}/src/app/util/process-cluster-message.c",
+    "${chip_root}/src/app/util/process-cluster-message.cpp",
     "${chip_root}/src/app/util/process-global-message.cpp",
     "${chip_root}/src/app/util/util.cpp",
     "${examples_plat_dir}/${efr32_family}/${efr32_board}/hal-config.h",
