+++ examples/lock-app/lock-common/BUILD.gn
@@ -25,18 +25,18 @@ config("includes") {
 source_set("lock-common") {
   sources = [
     "${chip_root}/examples/common/chip-app-server/DataModelHandler.cpp",
-    "${chip_root}/src/app/clusters/on-off-server/on-off.c",
+    "${chip_root}/src/app/clusters/on-off-server/on-off.cpp",
     "${chip_root}/src/app/util/af-event.cpp",
     "${chip_root}/src/app/util/af-main-common.cpp",
-    "${chip_root}/src/app/util/attribute-size.c",
-    "${chip_root}/src/app/util/attribute-storage.c",
-    "${chip_root}/src/app/util/attribute-table.c",
+    "${chip_root}/src/app/util/attribute-size.cpp",
+    "${chip_root}/src/app/util/attribute-storage.cpp",
+    "${chip_root}/src/app/util/attribute-table.cpp",
     "${chip_root}/src/app/util/binding-table.cpp",
     "${chip_root}/src/app/util/chip-message-send.cpp",
-    "${chip_root}/src/app/util/client-api.c",
+    "${chip_root}/src/app/util/client-api.cpp",
     "${chip_root}/src/app/util/ember-print.cpp",
     "${chip_root}/src/app/util/message.cpp",
-    "${chip_root}/src/app/util/process-cluster-message.c",
+    "${chip_root}/src/app/util/process-cluster-message.cpp",
     "${chip_root}/src/app/util/process-global-message.cpp",
     "${chip_root}/src/app/util/util.cpp",
     "gen/call-command-handler.c",
