+++ src/app/clusters/door-lock-server/door-lock-server-schedule.cpp
@@ -294,6 +294,7 @@ bool emberAfDoorLockClusterSetHolidayScheduleCallback(uint8_t holidayScheduleId,
     {
         holidayScheduleTable[holidayScheduleId].localStartTime = localStartTime;
         holidayScheduleTable[holidayScheduleId].localEndTime   = localEndTime;
+        // TODO: This cast may not be safe.  https://github.com/project-chip/connectedhomeip/issues/3578
         holidayScheduleTable[holidayScheduleId].operatingModeDuringHoliday =
             static_cast<EmberAfDoorLockOperatingMode>(operatingModeDuringHoliday);
         holidayScheduleTable[holidayScheduleId].inUse = true;
