+++ src/app/clusters/door-lock-server/door-lock-server-user.cpp
@@ -170,8 +170,10 @@ static uint8_t setUser(uint16_t userId, uint8_t userStatus, uint8_t userType, ui
     if (validCodeLength && userId < userTableSize)
     {
         EmberAfPluginDoorLockServerUser * user = &userTable[userId];
-        user->status                           = static_cast<EmberAfDoorLockUserStatus>(userStatus);
-        user->type                             = static_cast<EmberAfDoorLockUserType>(userType);
+        // TODO: Need to check validity.  https://github.com/project-chip/connectedhomeip/issues/3579
+        user->status = static_cast<EmberAfDoorLockUserStatus>(userStatus);
+        // TODO: Need to check validity.  https://github.com/project-chip/connectedhomeip/issues/3580
+        user->type = static_cast<EmberAfDoorLockUserType>(userType);
         memmove(user->code.rfid, code,
                 emberAfStringLength(code) + 1); // + 1 for Zigbee string length byte
 
@@ -229,6 +231,7 @@ bool emberAfDoorLockClusterGetUserTypeCallback(uint16_t userId)
 
 bool emberAfDoorLockClusterSetUserTypeCallback(uint16_t userId, uint8_t userType)
 {
+    // TODO: Need to validate userType.  https://github.com/project-chip/connectedhomeip/issues/3580
     uint8_t status = (emAfPluginDoorLockServerSetPinUserType(userId, static_cast<EmberAfDoorLockUserType>(userType))
                           ? 0x00   // success (per 7.3.2.17.21)
                           : 0x01); // failure (per 7.3.2.17.21)
