+++ src/app/util/ember-print.h
@@ -22,6 +22,10 @@
 #include <stdbool.h>
 #include <stdint.h>
 
+#ifdef __cplusplus
+extern "C" {
+#endif // __cplusplus
+
 /**
  * @brief Prints a log message
  * @param category - Currently ignored as zcl categories do not map to chip categories. Defaults to kLogCategory_Progress
@@ -51,5 +55,9 @@ void emberAfPrintBuffer(int category, const uint8_t * buffer, uint16_t length, b
  * @param string - Buffer to print as a string
  * */
 void emberAfPrintString(int category, const uint8_t * string);
+
+#ifdef __cplusplus
+} // extern "C"
+#endif // __cplusplus
 
 #endif // EMBER_PRINT_H
