+++ src/app/util/esi-management.cpp
